refactor(runner): extract session-events.ts from sandbox_agent.ts
Move the two session-listener routers (routeSessionEventToActiveTurn,
routePermissionRequestToActiveTurn) into session-events.ts. Pure code move
(today's versions verbatim, including the any-typed event/req params); monolith
imports them. Typecheck clean; 1158 runner unit tests pass.
